Improved content to have clear steps
We had feedback that this page and comms did not align and the comms was clearer for users by setting out steps.

I've tweaked the content to add similar steps e.g. Step 1: do this, Step 2: do this.

I think I've got the heading and their classes right but worth checking.