Example expandable error banner for discussion
In testing on production. If a CSV file structure error occurs, it is displayed as a long, illegible string.

Following discussions, we decided to implement an error notification with a details component to maintain a clean yet descriptive format. This is a quick mockup of how it might look, ready for further discussion within the team.

Signed-off-by: Andrew Scrivener <andrew.scrivener@education.gov.uk>